Split out "Descriptive Statistics" section ...
and move moment statistics items here.
