--- mathics/builtin/moments/__init__.py
@@ -1,8 +1,0 @@
-"""
-Statistics, Moments, and Generating Functions
-
-Moments or combinations of moments are used to summarize a distribution or data. Mean is used to indicate a center location, variance and standard deviation are used to indicate dispersion and covariance, and correlation to indicate dependence.
-"""
-
-# This tells documentation how to sort this module
-sort_order = "mathics.builtin.statistics-moments-and-generating-functions"
