+++ mathics/builtin/statistics/__init__.py
@@ -1,0 +1,9 @@
+"""
+Descriptive Statistics
+
+Function which operate on explicit data and symbolic representations of statistical distributions.
+"""
+
+# This tells documentation how to sort this module
+# Here we are also hiding "statistics" since this erroneously appears at the top level.
+sort_order = "mathics.builtin.descriptive-statistics"
