+++ mathics/builtin/structure.py
@@ -52,70 +52,6 @@ SymbolOperate = Symbol("Operate")
 SymbolSortBy = Symbol("SortBy")
 
 
-class Sort(Builtin):
-    """
-    <dl>
-      <dt>'Sort[$list$]'
-      <dd>sorts $list$ (or the elements of any other expression) according to canonical ordering.
-
-      <dt>'Sort[$list$, $p$]'
-      <dd>sorts using $p$ to determine the order of two elements.
-    </dl>
-
-    >> Sort[{4, 1.0, a, 3+I}]
-     = {1., 3 + I, 4, a}
-
-    Sort uses 'OrderedQ' to determine ordering by default.
-    You can sort patterns according to their precedence using 'PatternsOrderedQ':
-    >> Sort[{items___, item_, OptionsPattern[], item_symbol, item_?test}, PatternsOrderedQ]
-     = {item_symbol, item_ ? test, item_, items___, OptionsPattern[]}
-
-    When sorting patterns, values of atoms do not matter:
-    >> Sort[{a, b/;t}, PatternsOrderedQ]
-     = {b /; t, a}
-    >> Sort[{2+c_, 1+b__}, PatternsOrderedQ]
-     = {2 + c_, 1 + b__}
-    >> Sort[{x_ + n_*y_, x_ + y_}, PatternsOrderedQ]
-     = {x_ + n_ y_, x_ + y_}
-
-    #> Sort[{x_, y_}, PatternsOrderedQ]
-     = {x_, y_}
-    """
-
-    summary_text = "sort lexicographically or with any comparison function"
-
-    def apply(self, list, evaluation):
-        "Sort[list_]"
-
-        if isinstance(list, Atom):
-            evaluation.message("Sort", "normal")
-        else:
-            new_elements = sorted(list.elements)
-            return list.restructure(list.head, new_elements, evaluation)
-
-    def apply_predicate(self, list, p, evaluation):
-        "Sort[list_, p_]"
-
-        if isinstance(list, Atom):
-            evaluation.message("Sort", "normal")
-        else:
-
-            class Key:
-                def __init__(self, element):
-                    self.element = element
-
-                def __gt__(self, other):
-                    return (
-                        not Expression(p, self.element, other.element).evaluate(
-                            evaluation
-                        )
-                        is SymbolTrue
-                    )
-
-            new_elements = sorted(list.elements, key=Key)
-            return list.restructure(list.head, new_elements, evaluation)
-
-
 class SortBy(Builtin):
     """
     <dl>
