+++ setup.py
@@ -175,18 +175,18 @@ setup(
         "mathics.builtin.compile",
         "mathics.builtin.distance",
         "mathics.builtin.drawing",
-        "mathics.builtin.list",
         "mathics.builtin.fileformats",
         "mathics.builtin.files_io",
         "mathics.builtin.intfns",
+        "mathics.builtin.list",
         "mathics.builtin.matrices",
-        "mathics.builtin.moments",
         "mathics.builtin.numbers",
         "mathics.builtin.numpy_utils",
         "mathics.builtin.pymimesniffer",
         "mathics.builtin.pympler",
         "mathics.builtin.scipy_utils",
         "mathics.builtin.specialfns",
+        "mathics.builtin.statistics",
         "mathics.builtin.string",
         "mathics.builtin.vectors",
         "mathics.doc",
